move all css to stylefolder
