+++ src/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import './App.css';
+import './style/App.css';
 import NavBar from './components/navbar';
 import Categories from './pages/categories';
 import Books from './pages/books';
